Moved simplify to its own module.
